app/requirements.txt
fastapi
uvicorn
pydantic
pyserial
numpy
python-dotenv
# opencv-python
Pillow
# torch
# torchvision
# torchaudio
